reorg into src folder; fix dependencies for iOS build; working web, ios builds
