--- frontend/cesconfig.json
@@ -1,35 +1,0 @@
-{
-  "cesVersion": "2.12.2",
-  "projectName": "whereisglinda",
-  "packages": [
-    {
-      "name": "react-navigation",
-      "type": "navigation",
-      "options": {
-        "type": "stack"
-      }
-    },
-    {
-      "name": "stylesheet",
-      "type": "styling"
-    }
-  ],
-  "flags": {
-    "noGit": false,
-    "noInstall": false,
-    "overwrite": false,
-    "importAlias": true,
-    "packageManager": "npm",
-    "eas": false
-  },
-  "packageManager": {
-    "type": "npm",
-    "version": "10.8.1"
-  },
-  "os": {
-    "type": "Darwin",
-    "platform": "darwin",
-    "arch": "x64",
-    "kernelVersion": "23.6.0"
-  }
-}
