+++ frontend/gesture-handler.native.tsx
@@ -1,0 +1,2 @@
+// Only import react-native-gesture-handler on native platforms
+import 'react-native-gesture-handler';
