+++ frontend/gesture-handler.tsx
@@ -1,0 +1,1 @@
+// Don't import react-native-gesture-handler on web
