+++ frontend/package.json
@@ -1,47 +1,51 @@
 {
   "name": "whereisglinda",
   "version": "1.0.0",
+  "private": true,
+  "main": "index.js",
   "scripts": {
     "android": "expo run:android",
     "ios": "expo run:ios",
-    "start": "expo start",
+    "start": "expo start --dev-client",
     "lint": "eslint \"**/*.{js,jsx,ts,tsx}\" && prettier -c \"**/*.{js,jsx,ts,tsx,json}\"",
     "format": "eslint \"**/*.{js,jsx,ts,tsx}\" --fix && prettier \"**/*.{js,jsx,ts,tsx,json}\" --write",
     "web": "expo start --web"
   },
+  "overrides": {
+    "react": "18.3.1",
+    "react-dom": "18.3.1"
+  },
   "dependencies": {
-    "@expo/metro-runtime": "~4.0.1",
+    "@expo/metro-runtime": "~4.0.0",
     "@expo/vector-icons": "^14.0.4",
     "@react-native-community/geolocation": "^3.4.0",
+    "@react-native-masked-view/masked-view": "0.3.2",
     "@react-navigation/drawer": "^7.3.8",
-    "@react-navigation/native": "^7.0.19",
-    "expo": "~52.0.44",
+    "@react-navigation/elements": "^2.2.3",
+    "@react-navigation/native": "^7.0.12",
+    "@react-navigation/native-stack": "^7.1.13",
+    "expo": "~52.0.11",
+    "expo-asset": "~11.0.1",
     "expo-constants": "^17.0.0",
     "expo-location": "18.0.10",
-    "expo-system-ui": "~4.0.9",
+    "expo-module-scripts": "^4.0.5",
+    "expo-splash-screen": "~0.29.13",
+    "expo-system-ui": "~4.0.4",
     "react": "18.3.1",
     "react-dom": "18.3.1",
     "react-native": "0.76.9",
+    "react-native-edge-to-edge": "^1.1.3",
     "react-native-gesture-handler": "~2.20.2",
     "react-native-maps": "1.18.0",
-    "react-native-reanimated": "3.16.1",
+    "react-native-reanimated": "~3.16.1",
     "react-native-safe-area-context": "4.12.0",
     "react-native-screens": "~4.4.0",
     "react-native-sse": "^1.2.1",
     "react-native-web": "0.19.13",
     "react-native-web-google-api-maps": "github:chalda/react-native-web-google-api-maps"
   },
-  "overrides": {
-    "react-native-reanimated": "3.16.1",
-    "react": "18.3.1",
-    "react-dom": "18.3.1",
-    "react-native": "0.76.9",
-    "react-native-web": "0.19.13"
-  },
   "devDependencies": {
     "@babel/core": "^7.26.10",
-    "@babel/plugin-proposal-class-properties": "^7.18.6",
-    "@babel/plugin-proposal-decorators": "^7.25.9",
     "@babel/plugin-proposal-do-expressions": "^7.25.9",
     "@babel/plugin-proposal-export-default-from": "^7.25.9",
     "@babel/plugin-proposal-export-namespace-from": "^7.18.9",
@@ -52,32 +56,30 @@
     "@babel/plugin-proposal-optional-chaining": "^7.21.0",
     "@babel/preset-typescript": "^7.27.0",
     "@react-native-community/cli": "^18.0.0",
-    "@react-native/metro-config": "0.76.9",
+    "@react-native/js-polyfills": "^0.79.0",
+    "@react-native/metro-babel-transformer": "^0.79.0",
+    "@react-native/metro-config": "^0.79.0",
     "@react-navigation/devtools": "^7.0.25",
     "@types/node": "^22.14.0",
     "@types/react": "~18.3.20",
     "@types/react-native-web": "^0.19.0",
     "@typescript-eslint/eslint-plugin": "^8.29.0",
     "@typescript-eslint/parser": "^8.29.0",
-    "babel-plugin-module-resolver": "^5.0.2",
     "babel-plugin-react-native-web": "0.19.13",
     "dotenv": "^16.4.7",
     "eslint": "8.57.1",
     "eslint-config-universe": "^14.1.0",
     "expo-build-properties": "~0.13.2",
-    "expo-dev-client": "~5.0.20",
+    "expo-dev-client": "~5.0.4",
     "http-proxy-middleware": "^3.0.3",
-    "metro-react-native-babel-preset": "0.76.9",
+    "metro-config": "^0.82.1",
+    "metro-runtime": "^0.82.1",
     "prettier": "^3.5.3",
-    "react-native-paper": "^5.13.1",
     "ts-node": "^10.9.2",
     "typescript": "~5.8.3"
   },
   "eslintConfig": {
     "extends": "universe/native",
     "root": true
-  },
-  "main": "index.js",
-  "private": true,
-  "proxy": "http://localhost:8080"
+  }
 }
