--- frontend/setupProxy.js
@@ -1,15 +1,0 @@
-const { createProxyMiddleware } = require('http-proxy-middleware');
-
-module.exports = function (app) {
-  app.use(
-    '/api',
-    createProxyMiddleware({
-      target: 'http://localhost:8080',
-      changeOrigin: true,
-      pathRewrite: {
-        '^/api': '',
-      },
-      logLevel: 'debug',
-    })
-  );
-};
