+++ frontend/src/components/EditScreenInfo.tsx
