+++ frontend/src/components/MapContent.tsx
