+++ frontend/src/components/ScreenContent.tsx
