+++ frontend/src/screens/DriverLogin.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useContext } from 'react';
 import { View, Text, TextInput, Button, StyleSheet } from 'react-native';
 
-import { AppContext } from '../context/AppContext'; // Import AppContext for global state
+import { AppContext } from '../AppContext'; // Import AppContext for global state
 import { type RootStackNavigation } from '../navigation/RootStack';
 import { validateApiKey } from '../utils/api';
