+++ frontend/src/screens/HireGlinda.tsx
