--- frontend/utils/gesture-handler.ios.ts
@@ -1,1 +1,0 @@
-import 'react-native-gesture-handler';
